requirements.txt
django==2.0
django-rest-framework
django-rest-auth
django-allauth
django-cors-headers
